+++ example/ooni/dns_injection.cpp
@@ -35,12 +35,12 @@ int main(int argc, char **argv) {
         exit(1);
     }
 
-    mk::ooni::DnsInjectionTest()
+    mk::ooni::DnsInjection()
         .set_options("backend", backend)
         .set_options("geoip_country_path", "test/fixtures/GeoIP.dat")
         .set_options("geoip_asn_path", "test/fixtures/GeoIPASNum.dat")
         .set_verbosity(verbosity)
-        .set_input_file_path(argv[0])
+        .set_input_filepath(argv[0])
         .on_log([](uint32_t, const char *s) { std::cout << s << "\n"; })
         .run();
 }