+++ include/measurement_kit/ooni.hpp
@@ -2,8 +2,10 @@
 #ifndef MEASUREMENT_KIT_OONI_HPP
 #define MEASUREMENT_KIT_OONI_HPP
 #include <measurement_kit/ooni/collector_client.hpp>
-#include <measurement_kit/ooni/dns_injection_test.hpp>
+#include <measurement_kit/ooni/dns_injection.hpp>
 #include <measurement_kit/ooni/error.hpp>
-#include <measurement_kit/ooni/http_invalid_request_line_test.hpp>
-#include <measurement_kit/ooni/tcp_connect_test.hpp>
+#include <measurement_kit/ooni/http_invalid_request_line.hpp>
+#include <measurement_kit/ooni/ooni_test.hpp>
+#include <measurement_kit/ooni/tcp_connect.hpp>
+#include <measurement_kit/ooni/templates.hpp>
 #endif